requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.0.350
langchain-google-genai==0.0.5
python-dotenv==1.0.0
sqlalchemy>=1.4.42,<2.0
psycopg2-binary==2.9.9
alembic>=1.12.1
redis==5.0.1
celery==5.3.4
httpx==0.25.2
lxml==4.9.3
newspaper3k==0.2.8
nltk==3.8.1
slowapi==0.1.9
bleach==6.1.0
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
asyncpg==0.29.0
databases[postgresql]==0.5.5
